chore: test pulling progress bar logic into hook
